fix: Fix user assignees & clean up example yamls
